tests/wpt/web-platform-tests/webstorage/event_session_newvalue.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>WebStorage Test: sessionStorage event - newValue</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script>
      function fail(msg) {
          t.step(function() {
              assert_unreached(msg);
          });
          t.done();
      }
  </script>
 </head>
 <body>
    <h1>event_session_newValue</h1>
    <div id="log"></div>
    <script>
        var t = async_test("newvalue property test of session event");

        test(function() {
            sessionStorage.clear();
            var expected = ['user1', 'user2', null]
            function onStorageEvent(event) {
                t.step(function() {
                    assert_equals(event.newValue, expected.shift());
                });
                if (!expected.length) {
                    t.done();
                }
            }

            window.addEventListener('storage', onStorageEvent, false);

            var el = document.createElement("iframe");
            el.setAttribute('id', 'ifrm');
            el.setAttribute('src', 'iframe/session_change_item_iframe.html');
            document.body.appendChild(el);
        }, "Session event is fired due to an invocation of the setItem(), clear() methods.");
    </script>
 </body>
</html>
